feat(threaded-replies): Move Replies to BaseComment
