+++ i18n/en-US.properties
@@ -170,10 +170,6 @@ be.contentSidebar.activityFeed.comment.commentPostedFullDateTime = {time, date, 
 be.contentSidebar.activityFeed.comment.commentResolveMenuItem = Resolve
 # Text to show on menu item to unresolve the comment
 be.contentSidebar.activityFeed.comment.commentUnresolveMenuItem = Unresolve
-# Text to show on button to start replying to comment
-be.contentSidebar.activityFeed.comment.reply = Reply
-# Text to show on reply form input placeholder
-be.contentSidebar.activityFeed.comment.replyInThread = Reply in thread
 # Placeholder for approvers input
 be.contentSidebar.activityFeed.commentForm.approvalAddAssignee = Add an assignee
 # Label for checkbox to add approvers to a comment
