--- src/elements/content-sidebar/activity-feed/comment/Replies.js
@@ -1,92 +1,0 @@
-// @flow
-import * as React from 'react';
-import noop from 'lodash/noop';
-
-import BaseComment from './BaseComment';
-import LoadingIndicator from '../../../../components/loading-indicator';
-
-import type { GetAvatarUrlCallback, GetProfileUrlCallback } from '../../../common/flowTypes';
-import type { Translations } from '../../flowTypes';
-import type { SelectorItems, User } from '../../../../common/types/core';
-import type { BoxCommentPermission, Comment as CommentType } from '../../../../common/types/feed';
-
-import './Replies.scss';
-
-type ReplyProps = {
-    hasReplies?: boolean,
-    isRepliesLoading?: boolean,
-    onReplySelect: (isSelected: boolean) => void,
-    replies?: CommentType[],
-};
-
-type Props = ReplyProps & {
-    currentUser?: User,
-    getAvatarUrl: GetAvatarUrlCallback,
-    getMentionWithQuery?: (searchStr: string) => void,
-    getUserProfileUrl?: GetProfileUrlCallback,
-    mentionSelectorContacts?: SelectorItems<>,
-    translations?: Translations,
-};
-
-const Replies = ({
-    currentUser,
-    getAvatarUrl,
-    getMentionWithQuery,
-    getUserProfileUrl,
-    hasReplies = false,
-    isRepliesLoading = false,
-    mentionSelectorContacts,
-    onReplySelect,
-    replies = [],
-    translations,
-}: Props) => {
-    const getReplyPermissions = (reply: CommentType): BoxCommentPermission => {
-        const { permissions: { can_delete = false, can_edit = false, can_resolve = false } = {} } = reply;
-        return {
-            can_delete,
-            can_edit,
-            can_resolve,
-        };
-    };
-
-    if (!hasReplies) {
-        return null;
-    }
-
-    return (
-        <div className="bcs-Replies">
-            <div className="bcs-Replies-content">
-                {isRepliesLoading && (
-                    <div className="bcs-Replies-loading" data-testid="replies-loading">
-                        <LoadingIndicator />
-                    </div>
-                )}
-                <ol className="bcs-Replies-list">
-                    {replies.map(reply => {
-                        const { id, type } = reply;
-
-                        return (
-                            <li key={`${type}${id}`}>
-                                <BaseComment
-                                    {...reply}
-                                    currentUser={currentUser}
-                                    getAvatarUrl={getAvatarUrl}
-                                    getMentionWithQuery={getMentionWithQuery}
-                                    getUserProfileUrl={getUserProfileUrl}
-                                    mentionSelectorContacts={mentionSelectorContacts}
-                                    onSelect={onReplySelect}
-                                    onDelete={noop}
-                                    onEdit={noop}
-                                    permissions={getReplyPermissions(reply)}
-                                    translations={translations}
-                                />
-                            </li>
-                        );
-                    })}
-                </ol>
-            </div>
-        </div>
-    );
-};
-
-export default Replies;
